refactor(discord): split commands.ts into command-group modules
commands.ts was 1146 lines — slash-command defs, the interactionCreate
router, and ~16 per-command handlers in one file. Split into a commands/
folder mirroring the telegram layout:
  - definitions — SlashCommandBuilder defs + per-guild/global registration
  - shared      — reply / chatId / client helpers
  - router      — interactionCreate listener + slash-command dispatch
  - info        — /start /help /memory /ping /plugins
  - session     — /reset /status
  - settings    — /model /effort /pulse /settings
  - admin       — /restart /metrics /dream /admin

index.ts re-exports registerCommandsForGuilds + registerInteractionRouter.
Importer updated to the barrel path.
