+++ src/frontend/discord/commands/info.ts
@@ -1,0 +1,135 @@
+/**
+ * Informational commands — /start, /help, /memory, /ping, /plugins.
+ */
+
+import { readFileSync, existsSync } from "node:fs";
+import {
+  type ChatInputCommandInteraction,
+  type Client,
+  MessageFlags,
+} from "discord.js";
+import { files } from "../../../util/paths.js";
+import { formatDuration } from "../helpers.js";
+import { getLoadedPlugins } from "../../../core/plugin.js";
+import { DISCORD_MAX_TEXT, DISCORD_SAFE_RESERVE } from "../formatting.js";
+import { reply } from "./shared.js";
+
+export async function handleStart(
+  i: ChatInputCommandInteraction,
+): Promise<void> {
+  await reply(
+    i,
+    [
+      "**🦅 Talon**",
+      "",
+      "Agentic AI harness for Discord.",
+      "",
+      "Mention me, reply to me, or DM me. Attach files, photos, voice notes — I'll read them.",
+      "",
+      "/status  /reset  /help",
+    ].join("\n"),
+  );
+}
+
+export async function handleHelp(
+  i: ChatInputCommandInteraction,
+  c: Client,
+): Promise<void> {
+  const botMention = c.user ? `<@${c.user.id}>` : "@bot";
+  await reply(
+    i,
+    [
+      "**🦅 Talon — Help**",
+      "",
+      "**Settings**",
+      "  /settings — view and change all chat settings",
+      "  /model — show or change model",
+      "  /effort — set thinking effort (off, low, medium, high, max)",
+      "  /pulse — toggle periodic check-ins (on/off)",
+      "",
+      "**Session**",
+      "  /status — session info, usage, and stats",
+      "  /metrics — aggregate performance metrics (admin)",
+      "  /memory — view what Talon remembers",
+      "  /dream — force memory consolidation now (admin)",
+      "  /ping — health check with latency",
+      "  /reset — clear session and start fresh",
+      "  /restart — restart the bot process (admin)",
+      "  /plugins — list loaded plugins",
+      "  /help — this message",
+      "",
+      "**Input**",
+      "  Text, photos, documents, voice notes, audio, videos, GIFs, replies, attachments.",
+      "",
+      "**Servers**",
+      `  In a server, mention ${botMention} or reply to me to activate.`,
+      "",
+      "**DM**",
+      "  Whitelisted users can DM me directly.",
+    ].join("\n"),
+    true,
+  );
+}
+
+export async function handleMemory(
+  i: ChatInputCommandInteraction,
+): Promise<void> {
+  try {
+    const memoryPath = files.memory;
+    if (!existsSync(memoryPath)) {
+      await reply(
+        i,
+        "No memory file yet. I'll create one as I learn about you.",
+        true,
+      );
+      return;
+    }
+    const content = readFileSync(memoryPath, "utf-8").trim();
+    if (!content) {
+      await reply(i, "Memory file is empty. I'll update it as we chat.", true);
+      return;
+    }
+    const budget = DISCORD_MAX_TEXT - DISCORD_SAFE_RESERVE;
+    const display =
+      content.length > budget
+        ? content.slice(0, budget) + "\n\n... (truncated)"
+        : content;
+    await reply(i, display, true);
+  } catch {
+    await reply(i, "Could not read memory file.", true);
+  }
+}
+
+export async function handlePing(
+  i: ChatInputCommandInteraction,
+): Promise<void> {
+  const start = Date.now();
+  await i.deferReply({ flags: MessageFlags.Ephemeral });
+  const apiLatency = Date.now() - start;
+  // client.ws.ping is the WebSocket heartbeat RTT (the real gateway latency).
+  const wsPing = i.client.ws.ping;
+  const uptime = formatDuration(process.uptime() * 1000);
+  await i.editReply(
+    `Pong! WS: ${wsPing}ms · REST: ${apiLatency}ms\nGateway: ✓ | Uptime: ${uptime}`,
+  );
+}
+
+export async function handlePlugins(
+  i: ChatInputCommandInteraction,
+): Promise<void> {
+  const plugins = getLoadedPlugins();
+  if (plugins.length === 0) {
+    await reply(i, "No plugins loaded.", true);
+    return;
+  }
+  const lines = plugins.map((p) => {
+    const ver = p.plugin.version ? ` v${p.plugin.version}` : "";
+    const desc = p.plugin.description ? ` — ${p.plugin.description}` : "";
+    const mcp = p.plugin.mcpServerPath ? " [MCP]" : "";
+    const fe = p.plugin.frontends?.length
+      ? ` (${p.plugin.frontends.join(", ")})`
+      : "";
+    return `• **${p.plugin.name}**${ver}${mcp}${fe}${desc}`;
+  });
+  await reply(i, `**Plugins (${plugins.length})**\n\n${lines.join("\n")}`, true);
+}
